+++ app/playlist_ui/playlistcontroller.h
@@ -19,6 +19,7 @@
 #include <QHash>
 
 class QDropEvent;
+class TracksMimeData;
 
 namespace PlaylistUi {
   class Controller : public QObject {
@@ -32,6 +33,7 @@ namespace PlaylistUi {
     void changed(const std::shared_ptr<Playlist::Playlist> pl);
     void durationOfSelectedChanged(quint32 total_duration);
     void createPlaylistRequested(const QList<QDir> &filepaths, const QString &libraryDir);
+    void createPlaylistFromTracksRequested(const QVector<Track> &tracks, const QString &name);
 
   public slots:
     void on_load(const std::shared_ptr<Playlist::Playlist> pi);
@@ -43,6 +45,7 @@ namespace PlaylistUi {
     void on_trackMetaChanged(const Track &t);
     void on_appendToPlaylist(const QList<QDir> &filepaths);
     void on_appendTracks(const QVector<Track> &tracks);
+    void on_removeTracks(quint64 playlist_uid, const QVector<Track> &tracks);
     void sortBy(const QString &criteria);
 
   private slots:
@@ -70,8 +73,9 @@ namespace PlaylistUi {
     void updateStreamSpans();
     void eventFilterTableView(QEvent *event);
     void eventFilterViewport(QEvent *event);
-    bool handleExternalDnd(QEvent *event);
-    void onExternalDrop(QDropEvent *event);
+    bool handleDnd(QEvent *event);
+    void onDrop(QDropEvent *event);
+    const TracksMimeData *droppedTracks(QDropEvent *event) const;
 
     void loadColumnsConfig();